refactor: Dao에서 발생시키는 예외의 상위 예외인 RepositoryException 생성
