--- src/main/java/wooteco/subway/exception/DuplicatedNameException.java
@@ -1,10 +1,0 @@
-package wooteco.subway.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class DuplicatedNameException extends ClientRuntimeException {
-
-    public DuplicatedNameException(final String message) {
-        super(message, HttpStatus.BAD_REQUEST);
-    }
-}
