+++ src/main/java/wooteco/subway/exception/repository/DataNotFoundException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.exception.repository;
+
+public class DataNotFoundException extends RepositoryException {
+
+    public DataNotFoundException(final String message) {
+        super(message);
+    }
+}
