+++ src/main/java/wooteco/subway/exception/repository/DuplicatedFieldException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.exception.repository;
+
+public class DuplicatedFieldException extends RepositoryException {
+
+    public DuplicatedFieldException(final String message) {
+        super(message);
+    }
+}
