+++ src/main/java/wooteco/subway/exception/repository/RepositoryException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.exception.repository;
+
+public class RepositoryException extends RuntimeException {
+
+    public RepositoryException(final String message) {
+        super(message);
+    }
+}
