+++ src/main/java/wooteco/subway/station/StationService.java
@@ -3,7 +3,7 @@ package wooteco.subway.station;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import wooteco.subway.exception.DataNotFoundException;
+import wooteco.subway.exception.repository.DataNotFoundException;
 
 @Service
 public class StationService {
