+++ src/test/java/wooteco/subway/line/section/SectionDaoTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import wooteco.subway.exception.DataNotFoundException;
+import wooteco.subway.exception.repository.DataNotFoundException;
 import wooteco.subway.line.Line;
 import wooteco.subway.line.LineDao;
